telepresence status: Move in to its own file
